First attempt at a simplification mechanism
Usage is given in the docstring for `simplify`.

The simplification proceeds by:

1. figuring out a minimal set of linearly independant units from the
   input quantity/units. The set of units are chosen to have the
   smallest possible dimensions at this stage.
1. projecting the input units onto this minimal set units. The set can
   be non-orthogonal, so the projection takes place via a iterating
   projection algorithm. It is not an exact algorithm, so some rounding
   is done at the end. In most cases, I would expect this is sufficient
   to recover the exact projection.
1. conversion to the new units using uconvert

The main problem is figuring out how to project onto a non-orthogonal
basis. There are probably other (better?) ways of doing this.

Another issue is how to figure out some euristics to choose the linearly
independent units. We could try and add conditions to favor SI units, and/or
units with the smallest prefix (as in smallest abs(log10(prefix))), and/or
even the units with the smallest screen foot-print when dumping to STDOUT.