Allow inversion of unitful matrices
At least in some cases, the pivot step in the matrix inversion requires
scaling one coefficient by another. However this fails when implemented
in terms of two operations: first an inversion which gives an object
with an inverse unit (rather than a dimensionless quantity), then a
`*=` multiplication. The second step tries (incorrectly) to change the
dimension of the matrix coefficient to dimensionless.

The relevant lines in `linalg/lu.jl` are:

```Julia
Akkinv = inv(A[k,k])
for i = k+1:m
    A[i,k] *= Akkinv
end
```

This commit hacks an `inv` function for the relevant matrices that first
strips the units from the matrices.