proof of concept for associating a system of units with modules
Default set of units now lives in `Unitful.Defaults` instead of being
installed as part of the build system

I've tried to keep the changes as minimal as possible to make the diff
easy to interpret. I also disabled one test (temporarily) because it
creates a new unit. The test needs to be re-written to create that unit
targeting some module.