WIP: Add DifferenceQuantities
These quantities represent differences between units.
At the moment, this is restricted to affine units, since
that is where this makes the most sense. However, in theory
this kind of thing could be used where-ever the scale is
non-absolute (e.g. angle differences) - though perhaps that
is best done by a AffineUnit with no offset for a particular
application.

This is by no means done, but I figured I'd put it up for
a discussion. I'm planning to play with this a bit and see
if I like it or not.

Demo:
```
julia> 1°C + 1Δ°C
2 °C

julia> 1°C + 1°C
ERROR: AffineError: an invalid operation was attempted with affine quantities: 1 °C + 1 °C
Stacktrace:
 [1] +(x::Quantity{Int64, 𝚯, Unitful.FreeUnits{(K,), 𝚯, Unitful.Affine{-5463//20}}}, y::Quantity{Int64, 𝚯, Unitful.FreeUnits{(K,), 𝚯, Unitful.Affine{-5463//20}}})
   @ Unitful ~/.julia/dev/Unitful/src/quantities.jl:184
 [2] top-level scope
   @ REPL[5]:1

julia> 5°C - 2°C
3 Δ°C
```